requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
attrs==24.3.0
ccxt==4.4.40
ccxtpro==1.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dateparser==1.2.0
distlib==0.3.9
filelock==3.16.1
frozenlist==1.5.0
identify==2.6.5
idna==3.10
multidict==6.1.0
nodeenv==1.9.1
platformdirs==4.3.6
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
PyQt6==6.8.0
pyqt6-multiselect-combobox==1.1.1
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qasync==0.27.1
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvloop==0.21.0
websockets==14.1
yarl==1.18.3
